Add a download speed tracker.

+++ server/RdtClient.Service/Services/DownloadManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -15,7 +16,7 @@ namespace RdtClient.Service.Services
 {
     public static class DownloadManager
     {
-        public static readonly ConcurrentDictionary<Guid, Download> ActiveDownloads = new ConcurrentDictionary<Guid, Download>();
+        public static readonly Dictionary<Guid, Download> ActiveDownloads = new Dictionary<Guid, Download>();
 
         static DownloadManager()
         {
@@ -33,10 +34,20 @@ namespace RdtClient.Service.Services
                 return;
             }
 
+            download.Progress = 0;
+            download.BytesLastUpdate = 0;
+            download.NextUpdate = DateTime.UtcNow.AddSeconds(1);
+            download.Speed = 0;
+
             var fileUrl = download.Link;
 
             var uri = new Uri(fileUrl);
             var filePath = Path.Combine(destinationFolderPath, uri.Segments.Last());
+
+            if (!Directory.Exists(destinationFolderPath))
+            {
+                Directory.CreateDirectory(destinationFolderPath);
+            }
 
             var webRequest = WebRequest.Create(fileUrl);
             webRequest.Method = "HEAD";
@@ -56,7 +67,6 @@ namespace RdtClient.Service.Services
             {
                 await using var stream = response.GetResponseStream();
                 await using var fileStream = new FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.Write);
-
                 var buffer = new Byte[4096];
 
                 while (fileStream.Length < response.ContentLength)
@@ -67,8 +77,17 @@ namespace RdtClient.Service.Services
                     {
                         fileStream.Write(buffer, 0, read);
 
-                        download.Progress = (Int32) (fileStream.Length * 100 / responseLength);
-                        ActiveDownloads.TryAdd(download.DownloadId, download);
+                        ActiveDownloads[download.DownloadId].Progress = (Int32) (fileStream.Length * 100 / responseLength);
+
+                        if (DateTime.UtcNow > ActiveDownloads[download.DownloadId]
+                            .NextUpdate)
+                        {
+                            ActiveDownloads[download.DownloadId].Speed = fileStream.Length - ActiveDownloads[download.DownloadId].BytesLastUpdate;
+                            ActiveDownloads[download.DownloadId].NextUpdate = DateTime.UtcNow.AddSeconds(1);
+                            ActiveDownloads[download.DownloadId].BytesLastUpdate = fileStream.Length;
+
+                            Debug.WriteLine($"{fileStream.Length}/{responseLength} {ActiveDownloads[download.DownloadId].Speed}");
+                        }
                     }
                     else
                     {
@@ -77,26 +96,48 @@ namespace RdtClient.Service.Services
                 }
             }
 
+            ActiveDownloads[download.DownloadId].Speed = 0;
+
             try
             {
-                await using Stream stream = File.OpenRead(filePath);
-
-                var reader = ReaderFactory.Open(stream);
-                while (reader.MoveToNextEntry())
+                if (filePath.EndsWith(".rar"))
                 {
-                    if (!reader.Entry.IsDirectory)
+                    await UpdateStatus(download.DownloadId, DownloadStatus.Unpacking, TorrentStatus.Downloading);
+
+                    await using (Stream stream = File.OpenRead(filePath))
                     {
-                        Console.WriteLine(reader.Entry.Key);
-                        reader.WriteEntryToDirectory(destinationFolderPath,
-                                                     new ExtractionOptions
-                                                     {
-                                                         ExtractFullPath = true,
-                                                         Overwrite = true
-                                                     });
+                        var reader = ReaderFactory.Open(stream);
+                        while (reader.MoveToNextEntry())
+                        {
+                            if (reader.Entry.IsDirectory)
+                            {
+                                continue;
+                            }
+
+                            reader.WriteEntryToDirectory(destinationFolderPath,
+                                                         new ExtractionOptions
+                                                         {
+                                                             ExtractFullPath = true,
+                                                             Overwrite = true
+                                                         });
+                        }
+                    }
+
+                    var retryCount = 0;
+                    while (File.Exists(filePath) && retryCount < 10)
+                    {
+                        retryCount++;
+
+                        try
+                        {
+                            File.Delete(filePath);
+                        }
+                        catch
+                        {
+                            await Task.Delay(1000);
+                        }
                     }
                 }
-
-                File.Delete(filePath);
             }
             catch
             {
@@ -105,7 +146,7 @@ namespace RdtClient.Service.Services
 
             await UpdateStatus(download.DownloadId, DownloadStatus.Finished, TorrentStatus.Finished);
 
-            ActiveDownloads.TryRemove(download.DownloadId, out _);
+            ActiveDownloads.Remove(download.DownloadId);
         }
 
         private static async Task UpdateStatus(Guid downloadId, DownloadStatus downloadStatus, TorrentStatus torrentStatus)
